Fix AppVeyor download link to "develop" artifacts
- Thanks to github:JasperGeurtz for mentioning that the AppVeyor
download link for the artifacts from the "develop" branch was pointing
to the latest build of any branch. E.g. if "master" was updated, it
would point to the latest "master" artifacts. If "raw" was updated, it
would point to the latest "raw" artifacts. The download link for the
"develop" branch should now point to the "develop" artifacts only.